LPS-74666 Move TestRelatedEntryIndexerRegistry into BaseIndexerGetFullQueryTest
