+++ portal-impl/test/unit/com/liferay/portal/kernel/search/BaseIndexerGetFullQueryTest.java
@@ -22,18 +22,25 @@ import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.util.Props;
 import com.liferay.portal.kernel.util.PropsUtil;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.search.test.TestIndexerRegistry;
-import com.liferay.portal.search.test.TestRelatedEntryIndexerRegistry;
 import com.liferay.portlet.documentlibrary.util.DLFileEntryIndexer;
 import com.liferay.portlet.messageboards.util.MBMessageIndexer;
 import com.liferay.registry.BasicRegistryImpl;
 import com.liferay.registry.Registry;
 import com.liferay.registry.RegistryUtil;
+import com.liferay.registry.ServiceReference;
+import com.liferay.registry.ServiceTracker;
+import com.liferay.registry.ServiceTrackerCustomizer;
 
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 import javax.portlet.PortletRequest;
 import javax.portlet.PortletResponse;
@@ -265,6 +272,131 @@ public class BaseIndexerGetFullQueryTest extends PowerMockito {
 
 		@Override
 		protected void doReindex(String[] ids) throws Exception {
+		}
+
+	}
+
+	private class TestRelatedEntryIndexerRegistry
+		implements RelatedEntryIndexerRegistry {
+
+		public TestRelatedEntryIndexerRegistry() {
+			Registry registry = RegistryUtil.getRegistry();
+
+			_serviceTracker = registry.trackServices(
+				RelatedEntryIndexer.class,
+				new RelatedEntryIndexerServiceTrackerCustomizer());
+
+			_serviceTracker.open();
+		}
+
+		public void destroy() {
+			if (_serviceTracker != null) {
+				_serviceTracker.close();
+			}
+
+			_serviceTracker = null;
+		}
+
+		@Override
+		public List<RelatedEntryIndexer> getRelatedEntryIndexers() {
+			List<RelatedEntryIndexer> relatedEntryIndexers = new ArrayList<>();
+
+			for (List<RelatedEntryIndexer> currentRelatedEntryIndexers :
+					_relatedEntryIndexers.values()) {
+
+				relatedEntryIndexers.addAll(currentRelatedEntryIndexers);
+			}
+
+			return relatedEntryIndexers;
+		}
+
+		@Override
+		public List<RelatedEntryIndexer> getRelatedEntryIndexers(Class clazz) {
+			return getRelatedEntryIndexers(clazz.getName());
+		}
+
+		@Override
+		public List<RelatedEntryIndexer> getRelatedEntryIndexers(
+			String className) {
+
+			List<RelatedEntryIndexer> relatedEntryIndexers =
+				_relatedEntryIndexers.get(className);
+
+			if (relatedEntryIndexers == null) {
+				relatedEntryIndexers = Collections.emptyList();
+			}
+
+			return relatedEntryIndexers;
+		}
+
+		private final Map<String, List<RelatedEntryIndexer>>
+			_relatedEntryIndexers = new ConcurrentHashMap<>();
+		private ServiceTracker<RelatedEntryIndexer, RelatedEntryIndexer>
+			_serviceTracker;
+
+		private class RelatedEntryIndexerServiceTrackerCustomizer
+			implements ServiceTrackerCustomizer
+				<RelatedEntryIndexer, RelatedEntryIndexer> {
+
+			@Override
+			public RelatedEntryIndexer addingService(
+				ServiceReference<RelatedEntryIndexer> serviceReference) {
+
+				Registry registry = RegistryUtil.getRegistry();
+
+				RelatedEntryIndexer relatedEntryIndexer = registry.getService(
+					serviceReference);
+
+				String relatedEntryIndexerClassName =
+					(String)serviceReference.getProperty(
+						"related.entry.indexer.class.name");
+
+				if (Validator.isNull(relatedEntryIndexerClassName)) {
+					throw new IllegalStateException(
+						"Service must contain a " +
+							"related.entry.indexer.class.name");
+				}
+
+				List<RelatedEntryIndexer> relatedEntryIndexers =
+					_relatedEntryIndexers.get(relatedEntryIndexerClassName);
+
+				if (relatedEntryIndexers == null) {
+					relatedEntryIndexers = new ArrayList<>();
+
+					_relatedEntryIndexers.put(
+						relatedEntryIndexerClassName, relatedEntryIndexers);
+				}
+
+				relatedEntryIndexers.add(relatedEntryIndexer);
+
+				return relatedEntryIndexer;
+			}
+
+			@Override
+			public void modifiedService(
+				ServiceReference<RelatedEntryIndexer> serviceReference,
+				RelatedEntryIndexer relatedEntryIndexer) {
+			}
+
+			@Override
+			public void removedService(
+				ServiceReference<RelatedEntryIndexer> serviceReference,
+				RelatedEntryIndexer relatedEntryIndexer) {
+
+				Registry registry = RegistryUtil.getRegistry();
+
+				registry.ungetService(serviceReference);
+
+				String relatedEntryIndexerClassName =
+					(String)serviceReference.getProperty(
+						"related.entry.indexer.class.name");
+
+				List<RelatedEntryIndexer> relatedEntryIndexers =
+					_relatedEntryIndexers.get(relatedEntryIndexerClassName);
+
+				relatedEntryIndexers.remove(relatedEntryIndexer);
+			}
+
 		}
 
 	}
