+++ portal-test-integration/src/com/liferay/portal/search/test/packageinfo
@@ -1,1 +1,1 @@
-version 3.1.0
+version 3.0.0
